Fixed to compile with VC++ 6.
git-svn-id: http://yate.null.ro/svn/yate/trunk@810 acf43c95-373e-0410-b603-e72c3f656dc1

+++ modules/rmanager.cpp
@@ -323,7 +323,7 @@ bool Connection::processLine(const char *line)
 	str.clear();
 	u_int32_t t = SysUsage::secRunTime();
 	if (m_machine) {
-	    str << "%%=uptime:" << t;
+	    str << "%%=uptime:" << (unsigned int)t;
 	    (str << ":").append(SysUsage::runTime(SysUsage::UserTime));
 	    (str << ":").append(SysUsage::runTime(SysUsage::KernelTime));
 	}
